feat: PrReviewRepository findById 추가 및 리뷰 수정 삭제 좋아요, 리뷰 수 count 설정 #10

+++ infra/repositories/prisma/PrReviewRepository.ts
@@ -6,12 +6,39 @@ export class PrReviewRepository implements ReviewRepository {
   private prisma = new PrismaClient();
 
   async create(review: CreateReviewDto) {
-    return this.prisma.review.create({
-      data: {
-        userId: review.userId,
-        codeId: review.codeId,
-        content: review.content,
-        parentId: review.parentId ?? null,
+    const [createdReview] = await this.prisma.$transaction([
+      this.prisma.review.create({
+        data: {
+          userId: review.userId,
+          codeId: review.codeId,
+          content: review.content,
+          parentId: review.parentId ?? null,
+        },
+      }),
+      this.prisma.user.update({
+        where: { id: review.userId },
+        data: { reviewCount: { increment: 1 } }, // 리뷰 작성 시 사용자 리뷰 수 증가
+      }),
+    ]);
+    return createdReview;
+  }
+
+  async findById(id: number) {
+    return this.prisma.review.findUnique({
+      where: { id },
+      include: {
+        user: {
+          select: {
+            id: true,
+            nickname: true,
+            imageUrl: true,
+            grade: {
+              select: {
+                name: true,
+              },
+            },
+          },
+        },
       },
     });
   }
@@ -31,9 +58,26 @@ export class PrReviewRepository implements ReviewRepository {
   }
 
   async delete(id: number) {
-    await this.prisma.review.delete({
+    const review = await this.prisma.review.findUnique({
       where: { id },
+      select: { userId: true },
     });
+
+    if (!review) {
+      throw new Error('리뷰를 찾을 수 없습니다.');
+    }
+
+    this.prisma.$transaction([
+      this.prisma.review.delete({
+        where: { id },
+      }),
+      this.prisma.user.update({
+        where: { id: review.userId },
+        data: {
+          reviewCount: { decrement: 1 },
+        },
+      }),
+    ]);
   }
 
   async findAllByCodeId(codeId: number): Promise<ReviewView[]> {
